// generated/src/app/api-models/batch-message.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BatchMessage {
  seqMessageId: number;
  batchId: string;
  transactionId: number;
  additionalKey: number;
  messageId: number;
  messageType: string;
  messageText: string;
  keyId1: string;
  keyId2: string;
  keyId3: string;
  keyId4: string;
  keyId5: string;
  keyId6: string;
  keyId7: string;
  keyId8: string;
  keyId9: string;
  keyId10: string;
  keyId11: string;
  keyId12: string;
  keyId13: string;
  keyId14: string;
  keyId15: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}